# Make scatter_* kernels to be multiple thread and gain 8x speedup.