--- a/tensorflow/core/kernels/scatter_functor.h
+++ b/tensorflow/core/kernels/scatter_functor.h
@@ -17,14 +17,16 @@ limitations under the License.
 #define TENSORFLOW_KERNELS_SCATTER_FUNCTOR_H_
 
 #include <type_traits>
+#include <atomic>
 
 #include "tensorflow/core/framework/tensor.h"
 #include "tensorflow/core/kernels/bounds_check.h"
 #include "tensorflow/core/platform/types.h"
+#include "tensorflow/core/util/work_sharder.h"
+#include "tensorflow/core/framework/op_kernel.h"
 
 namespace tensorflow {
 
-class OpKernelContext;
 typedef Eigen::ThreadPoolDevice CPUDevice;
 typedef Eigen::GpuDevice GPUDevice;
 #ifdef TENSORFLOW_USE_SYCL
@@ -140,17 +142,36 @@ struct ScatterFunctorBase {
     // indices and params sizes were validated in DoCompute().
     const Index N = static_cast<Index>(indices.size());
     const Index limit = static_cast<Index>(params.dimension(0));
-    for (Index i = 0; i < N; i++) {
-      // Grab the index and check its validity.  An earlier version of the
-      // code checked it and then grabbed it from memory a second time, which
-      // was a security risk since it could have changed in between.
-      const Index index = ::tensorflow::internal::SubtleMustCopy(indices(i));
-      if (!FastBoundsCheck(index, limit)) return i;
-      // Copy last Ndim-1 dimensions of updates[i] to params[index]
-      scatter_op::internal::Assign<op>::Run(params.template chip<0>(index),
-                                            updates.template chip<0>(i));
-    }
-    return -1;
+    auto worker_threads = c->device()->tensorflow_cpu_worker_threads();
+    auto locks = std::unique_ptr<std::atomic_flag[]>(
+            new std::atomic_flag[N]{ATOMIC_FLAG_INIT});
+    std::atomic_flag* locks_ptr = locks.get();
+    mutex mu;
+    Index result = -1 GUARDED_BY(mu);
+    auto work = [&params, &updates, &indices, &locks_ptr, N, limit, &mu, &result]
+            (int64 start, int64 end) {
+      Index mu_idx;
+      for (Index i = start; i < end; i++) {
+        // Grab the index and check its validity.  An earlier version of the
+        // code checked it and then grabbed it from memory a second time, which
+        // was a security risk since it could have changed in between.
+        const Index index = ::tensorflow::internal::SubtleMustCopy(indices(i));
+        if (!FastBoundsCheck(index, limit)) {
+          mutex_lock l(mu);
+          result = i;
+          return;
+        }
+        mu_idx = index % N;
+        while (locks_ptr[mu_idx].test_and_set(std::memory_order_acquire)) ; // spin
+        // Copy last Ndim-1 dimensions of updates[i] to params[index]
+        scatter_op::internal::Assign<op>::Run(params.template chip<0>(index),
+                                              updates.template chip<0>(i));
+        locks_ptr[mu_idx].clear(std::memory_order_release);
+      }
+    };
+    Shard(worker_threads->num_threads, worker_threads->workers, N,
+          updates.dimension(1) * sizeof(T), work);
+    return result;
   }
 };