--- a/tensorflow/core/kernels/scatter_functor.h
+++ b/tensorflow/core/kernels/scatter_functor.h
@@ -133,9 +133,9 @@ struct ScatterFunctor {
                    typename TTypes<Index>::ConstFlat indices);
 };
 
-template <typename Device, typename T, typename Index, scatter_op::UpdateOp op>
-struct ScatterFunctorBase {
-  Index operator()(OpKernelContext* c, const Device& d,
+template <typename T, typename Index, scatter_op::UpdateOp op>
+struct ScatterFunctor<CPUDevice, T, Index, op> {
+  Index operator()(OpKernelContext* c, const CPUDevice& d,
                    typename TTypes<T>::Matrix params,
                    typename TTypes<T>::ConstMatrix updates,
                    typename TTypes<Index>::ConstFlat indices) {
@@ -144,8 +144,10 @@ struct ScatterFunctorBase {
     const Index limit = static_cast<Index>(params.dimension(0));
     auto worker_threads = c->device()->tensorflow_cpu_worker_threads();
 
-    // Calculate the max number of flags.
+    // Limit the max number of flags in case of too much indices.
     // Based on the shard strategy in work_sharder.cc to limit the flags' size.
+    // Assume each cost unit ues one flag, and the minCostPerShard is 10000,
+    //   so limit the factor to 10000.
     static const int64 maxFactorPerThread = 10000;
     Index flags_size = static_cast<Index>(maxFactorPerThread * worker_threads->num_threads);
     flags_size = N <= flags_size ? N : flags_size;
@@ -158,25 +160,55 @@ struct ScatterFunctorBase {
     Index result = -1 GUARDED_BY(mu);
     auto work = [&params, &updates, &indices, &flags, flags_size, limit, &mu, &result]
             (int64 start, int64 end) {
+
+#define BOUNDS_CHECK                                                    \
+      if (!FastBoundsCheck(index, limit)) {                             \
+        mutex_lock l(mu);                                               \
+        result = i;                                                     \
+        return;                                                         \
+      }
+
+#define ACQUIRE_LOCK                                                   \
+      flags_idx = index % flags_size;                                  \
+      while (flags[flags_idx].test_and_set(std::memory_order_acquire)) ;
+
+
+#define RELEASE_LOCK  flags[flags_idx].clear(std::memory_order_release);
+
       Index flags_idx;
-      for (Index i = start; i < end; i++) {
-        // Grab the index and check its validity.  An earlier version of the
-        // code checked it and then grabbed it from memory a second time, which
-        // was a security risk since it could have changed in between.
-        const Index index = ::tensorflow::internal::SubtleMustCopy(indices(i));
-        if (!FastBoundsCheck(index, limit)) {
-          mutex_lock l(mu);
-          result = i;
-          return;
+      if (op != scatter_op::UpdateOp::ASSIGN || std::is_same<T, string>::value) {
+        for (Index i = start; i < end; i++) {
+          // Grab the index and check its validity.  An earlier version of the
+          // code checked it and then grabbed it from memory a second time, which
+          // was a security risk since it could have changed in between.
+          const Index index = ::tensorflow::internal::SubtleMustCopy(indices(i));
+          BOUNDS_CHECK
+          ACQUIRE_LOCK
+          // Copy last Ndim-1 dimensions of updates[i] to params[index]
+          scatter_op::internal::Assign<op>::Run(params.template chip<0>(index),
+                                                updates.template chip<0>(i));
+          RELEASE_LOCK
         }
-        flags_idx = index % flags_size;
-        while (flags[flags_idx].test_and_set(std::memory_order_acquire)) ; // spin
-        // Copy last Ndim-1 dimensions of updates[i] to params[index]
-        scatter_op::internal::Assign<op>::Run(params.template chip<0>(index),
-                                              updates.template chip<0>(i));
-        flags[flags_idx].clear(std::memory_order_release);
+      } else {
+        for (Index i = start; i < end; i++) {
+          // Grab the index and check its validity.  An earlier version of the
+          // code checked it and then grabbed it from memory a second time, which
+          // was a security risk since it could have changed in between.
+          const Index index = ::tensorflow::internal::SubtleMustCopy(indices(i));
+
+          BOUNDS_CHECK
+          ACQUIRE_LOCK
+          memmove(params.data() + index * params.dimension(1),
+                  updates.data() + i * updates.dimension(1),
+                  updates.dimension(1) * sizeof(T));
+          RELEASE_LOCK
+        }
       }
     };
+#undef RELEASE_LOCK
+#undef ACQUIRE_LOCK
+#undef BOUNDS_CHECK
+
     Shard(worker_threads->num_threads, worker_threads->workers, N,
           updates.dimension(1) * sizeof(T), work);
     return result;
@@ -207,83 +239,6 @@ struct ScatterFunctorBase <SYCLDevice, T, Index, op> {
   }
 };
 #endif // TENSORFLOW_USE_SYCL
-
-template <typename T, typename Index>
-struct ScatterFunctorBase<CPUDevice, T, Index, scatter_op::UpdateOp::ASSIGN> {
-  Index operator()(OpKernelContext* c, const CPUDevice& d,
-                   typename TTypes<T>::Matrix params,
-                   typename TTypes<T>::ConstMatrix updates,
-                   typename TTypes<Index>::ConstFlat indices) {
-    // indices and params sizes were validated in DoCompute().
-    const Index N = static_cast<Index>(indices.size());
-    const Index limit = static_cast<Index>(params.dimension(0));
-    auto worker_threads = c->device()->tensorflow_cpu_worker_threads();
-
-    // Calculate the max number of flags.
-    // Based on the shard strategy in work_sharder.cc to limit the flags' size.
-    static const int64 maxFactorPerThread = 10000;
-    Index flags_size = static_cast<Index>(maxFactorPerThread * worker_threads->num_threads);
-    flags_size = N <= flags_size ? N : flags_size;
-    // use atomic_flag and spin lock for multiple thread.
-    auto flags = std::unique_ptr<std::atomic_flag[]>(
-            new std::atomic_flag[flags_size]{ATOMIC_FLAG_INIT});
-
-    // store the result.
-    mutex mu;
-    Index result = -1 GUARDED_BY(mu);
-
-    auto work = [&params, &updates, &indices, &flags, flags_size, limit, &mu, &result]
-            (int64 start, int64 end) {
-      Index flags_idx;
-      if (!std::is_same<T, string>::value) {
-        for (Index i = start; i < end; i++) {
-          // Grab the index and check its validity.  An earlier version of the
-          // code checked it and then grabbed it from memory a second time, which
-          // was a security risk since it could have changed in between.
-          const Index index = ::tensorflow::internal::SubtleMustCopy(indices(i));
-          if (!FastBoundsCheck(index, limit)) {
-            mutex_lock l(mu);
-            result = i;
-            return;
-          }
-          flags_idx = index % flags_size;
-          // spin lock
-          while (flags[flags_idx].test_and_set(std::memory_order_acquire)) ;
-          memmove(params.data() + index * params.dimension(1),
-                  updates.data() + i * updates.dimension(1),
-                  updates.dimension(1) * sizeof(T));
-          flags[flags_idx].clear(std::memory_order_release);
-        }
-      } else {
-        for (Index i = start; i < end; i++) {
-          // Grab the index and check its validity.  An earlier version of the
-          // code checked it and then grabbed it from memory a second time, which
-          // was a security risk since it could have changed in between.
-          const Index index = ::tensorflow::internal::SubtleMustCopy(indices(i));
-          if (!FastBoundsCheck(index, limit)) {
-            mutex_lock l(mu);
-            result = i;
-            return;
-          }
-          flags_idx = index % flags_size;
-          // spin lock
-          while (flags[flags_idx].test_and_set(std::memory_order_acquire)) ;
-          // Copy last Ndim-1 dimensions of updates[i] to params[index]
-          scatter_op::internal::Assign<scatter_op::UpdateOp::ASSIGN>::Run(
-                  params.template chip<0>(index), updates.template chip<0>(i));
-          flags[flags_idx].clear(std::memory_order_release);
-        }
-      }
-    };
-    Shard(worker_threads->num_threads, worker_threads->workers, N,
-          updates.dimension(1) * sizeof(T), work);
-    return result;
-  }
-};
-
-template <typename T, typename Index, scatter_op::UpdateOp op>
-struct ScatterFunctor<CPUDevice, T, Index, op>
-        : ScatterFunctorBase<CPUDevice, T, Index, op>{};
 
 #ifdef TENSORFLOW_USE_SYCL
 template <typename T, typename Index, scatter_op::UpdateOp op>