Relate to [12358](https://github.com/tensorflow/tensorflow/issues/12358).
Use the lock to deal with the duplicate. After some test, the results show the atomic_flag and spin lock are the best choice and the performance is close  to the lock-free mode.
Benchmark with tfprof, the result shows about 8x speedup.
### Env
- **CPU**
32-cores
- **MEM**
126G
### Parameters
- ref: 10000000*100
- indices: 128000
- updates: 128000*100
```
# orignial
ScatterSub                    4400.00MB (65.69%, 6.66%),        39.50ms (43.19%, 9.00%),            0us (47.10%, 0.00%),        39.50ms (43.14%, 9.11%)
# lock-free
ScatterSub                    4400.00MB (65.69%, 6.66%),         3.84ms (50.64%, 2.20%),            0us (46.37%, 0.00%),         3.84ms (50.76%, 2.27%)
# with lock
ScatterSub                    4400.00MB (65.69%, 6.66%),         4.70ms (50.06%, 2.81%),            0us (45.62%, 0.00%),         4.70ms (50.19%, 2.90%)
```