@nolanliou 2017-08-25
@ekelsen code updated.