@ekelsen 2017-08-25
The code is much cleaner, thank you .  I think you could collapse the two loop bodies into one and pass the assignment function in.  Then no code is duplicated.  What do you think?