@nolanliou 2017-08-25
The strategy of string assignment is memcpy which is different with the other type. if collapse the two loop bodies into one, there shuold be a judgement inside the loop. maybe it's no so good.