@ekelsen 2017-08-25
If you pass the assignment op into the function there wouldn't be a conditional inside the loop.