@nolanliou 2017-09-04
@ekelsen Could you review the code if you have time?