@drpngx 2017-09-18
Here's the error:
```
FAIL: //tensorflow/python/kernel_tests:embedding_ops_test (shard 16 of 20) (see /var/lib/jenkins/workspace/tensorflow-pull-requests-cpu/bazel-ci_build-cache/.cache/bazel/_bazel_jenkins/eab0d61a99b6696edb3d2aff87b585e8/execroot/workspace/bazel-out/local-opt/testlogs/tensorflow/python/kernel_tests/embedding_ops_test/shard_16_of_20/test.log).
INFO: From Testing //tensorflow/python/kernel_tests:embedding_ops_test (shard 16 of 20):
==================== Test output for //tensorflow/python/kernel_tests:embedding_ops_test (shard 16 of 20):
2017-09-18 20:25:27.451882: I tensorflow/core/platform/cpu_feature_guard.cc:137] Your CPU supports instructions that this TensorFlow binary was not compiled to use: SSE4.1 SSE4.2 AVX AVX2 FMA
.F
======================================================================
FAIL: testRandom (__main__.ScatterAddSubTest)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/var/lib/jenkins/workspace/tensorflow-pull-requests-cpu/bazel-ci_build-cache/.cache/bazel/_bazel_jenkins/eab0d61a99b6696edb3d2aff87b585e8/execroot/workspace/bazel-out/local-opt/bin/tensorflow/python/kernel_tests/embedding_ops_test.runfiles/org_tensorflow/tensorflow/python/kernel_tests/embedding_ops_test.py", line 103, in testRandom
    self._TestCase(_AsLong(list(shape)), list(indices))
  File "/var/lib/jenkins/workspace/tensorflow-pull-requests-cpu/bazel-ci_build-cache/.cache/bazel/_bazel_jenkins/eab0d61a99b6696edb3d2aff87b585e8/execroot/workspace/bazel-out/local-opt/bin/tensorflow/python/kernel_tests/embedding_ops_test.runfiles/org_tensorflow/tensorflow/python/kernel_tests/embedding_ops_test.py", line 86, in _TestCase
    self.assertTrue(all((p_init == result).ravel()))
AssertionError: False is not true

----------------------------------------------------------------------
Ran 2 tests in 0.715s

FAILED (failures=1)
================================================================================
PASS: //tensorflow/python/kernel_tests:embedding_ops_test (shard 14 of 20)
PASS: //tensorflow/python/kernel_tests:gradient_correctness_test
PASS: //tensorflow/python/kernel_tests:embedding_ops_test (shard 15 of 20)
PASS: //tensorflow/python/kernel_tests:identity_op_py_test
PASS: //tensorflow/python/kernel_tests:identity_n_op_py_test
PASS: //tensorflow/python/kernel_tests:in_topk_op_test
PASS: //tensorflow/python/kernel_tests:io_ops_test
FAIL: //tensorflow/python/kernel_tests:embedding_ops_test (shard 17 of 20) (see /var/lib/jenkins/workspace/tensorflow-pull-requests-cpu/bazel-ci_build-cache/.cache/bazel/_bazel_jenkins/eab0d61a99b6696edb3d2aff87b585e8/execroot/workspace/bazel-out/local-opt/testlogs/tensorflow/python/kernel_tests/embedding_ops_test/shard_17_of_20/test.log).
INFO: From Testing //tensorflow/python/kernel_tests:embedding_ops_test (shard 17 of 20):
==================== Test output for //tensorflow/python/kernel_tests:embedding_ops_test (shard 17 of 20):
2017-09-18 20:25:27.213108: I tensorflow/core/platform/cpu_feature_guard.cc:137] Your CPU supports instructions that this TensorFlow binary was not compiled to use: SSE4.1 SSE4.2 AVX AVX2 FMA
.F
======================================================================
FAIL: testSubRandom (__main__.ScatterAddSubTest)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/var/lib/jenkins/workspace/tensorflow-pull-requests-cpu/bazel-ci_build-cache/.cache/bazel/_bazel_jenkins/eab0d61a99b6696edb3d2aff87b585e8/execroot/workspace/bazel-out/local-opt/bin/tensorflow/python/kernel_tests/embedding_ops_test.runfiles/org_tensorflow/tensorflow/python/kernel_tests/embedding_ops_test.py", line 110, in testSubRandom
    self._TestCase(_AsLong(list(shape)), list(indices), state_ops.scatter_sub)
  File "/var/lib/jenkins/workspace/tensorflow-pull-requests-cpu/bazel-ci_build-cache/.cache/bazel/_bazel_jenkins/eab0d61a99b6696edb3d2aff87b585e8/execroot/workspace/bazel-out/local-opt/bin/tensorflow/python/kernel_tests/embedding_ops_test.runfiles/org_tensorflow/tensorflow/python/kernel_tests/embedding_ops_test.py", line 86, in _TestCase
    self.assertTrue(all((p_init == result).ravel()))
AssertionError: False is not true

----------------------------------------------------------------------
Ran 2 tests in 2.722s
```

Note that it doesn't fail on all shards.