@nolanliou 2017-09-19
After debug, it turns out the float-point precision problem. multi thread leads to different calculation order which generate different-precision result.
The [stackoverflow question](https://stackoverflow.com/questions/7365790/how-can-floating-point-calculations-be-made-deterministic) explain that.
Should we take care of the calculation order to make sure the precision same all the time?