@ekelsen 2017-09-19
The documentation for the op does not guarantee ordering, but I would prefer the op to be deterministic if it can be done without much performance hit and memory overhead.