@nolanliou 2017-09-19
The extra calculation and memory usage is determined by the size of indices. Maybe it's not good to make the op deterministic.