@caisq 2017-09-19
@tensorflow-jenkins test this please