@nolanliou 2017-09-20
Use the assertAllClose with 1e-5 precision to check the result.