@drpngx 2017-09-20
Minor point: there is no need to set `rtol=0` AFAIK, I think it's a bit more idiomatic to omit that.