@nolanliou 2017-09-20
MacOS CPU Tests ERROR:
```
./tensorflow/core/kernels/scatter_functor.h:160:22: error: expected ';' at end of declaration
    Index result = -1 GUARDED_BY(mu);
```
what's going on?