@drpngx 2017-09-20
Timeout in `embedding_ops_test` on Windows, trying again.

Jenkins, test this please.