@nolanliou 2017-09-22
@drpngx Could you paste the error log about CPU test ?