@drpngx 2017-09-22
It's in `//tensorflow/cc/gradients:nn_grad_test`:
```
[RUN      ] NNGradTest.MaxPoolGradV2Helper
tensorflow/cc/gradients/nn_grad_test.cc:39: Failure
Expected: (max_error) < (2e-4), actual: 0.446481 vs 0.0002
[  FAILED  ] NNGradTest.MaxPoolGradV2Helper (3 ms)
```