@caisq 2017-09-22
@tensorflow-jenkins test this please.