@drpngx 2017-09-22
7 tests failed. Look for

```
19:25:01 292/297 Test #105: C:/tf_jenkins/home/workspace/tensorflow-pr-win-cmake-py/tensorflow/python/kernel_tests/embedding_ops_test.py ....................................***Timeout 600.28 sec
```
in the [log](https://ci.tensorflow.org/job/tensorflow-pr-win-cmake-py/4452/consoleFull).