@drpngx 2017-09-27
Actually, it looks like a lot of timeouts? If it's not due some flakiness, then it means that this op slows down significantly.