@nolanliou 2017-09-27
The linux CPU tests is all passed while the windows is timeout. I have changed the atomic_flag to mutex, please have another try.