@sb2nov 2017-10-01
@nolanliou can you rebase and check if the test failures are fixed, if not take a look at them. That would help us move the merge forward. Thanks for the help.