@nolanliou 2017-10-02
Rebase done. there are some problems in windows test, but I don't have accessable windows environment to do the test.